Popravljen HealthBar

+++ Assets/Scripts/PlayerHealth.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 public class PlayerHealth : MonoBehaviour
 {
 
-    public int maxHealth = 10;
-    public int currentHealth;
+    public float maxHealth = 10;
+    public float currentHealth;
 
     void Start()
     {
